Update patch set 1
Patch Set 1:

My concern is that you are causing the inheritance chain to become rather messy. As it is you are forcing two audio HALs to be defined. If you wish to do this, perhaps having the dual SIM klte devices stand separate from klte-common is the way to go.

Patch-set: 1